+++ packages/erd-generator/src/index.ts
@@ -1,2 +1,4 @@
 export { ERDGenerator } from './ERDGenerator';
+export { DataverseClient } from './DataverseClient';
 export * from './types';
+export type { DataverseConfig } from './DataverseClient';